@@ -1,12 +1,12 @@
 {
   "name": "@osjs/osjs",
-  "version": "3.0.5",
+  "version": "3.0.7",
   "description": "OS.js v3",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "eslint": "eslint src/client/**/*.js src/server/**/*.js",
-    "stylelint": "stylelint src/client/*.scss",
+    "test": "npm run eslint && npm run stylelint",
+    "eslint": "eslint 'src/client/**/*.js' 'src/server/**/*.js'",
+    "stylelint": "stylelint 'src/client/*.scss'",
     "serve": "node src/server/index.js",
     "build:manifest": "osjs-cli package:discover",
     "package:discover": "osjs-cli package:discover",
@@ -26,47 +26,47 @@
   "author": "Anders Evenrud <andersevenrud@gmail.com>",
   "license": "BSD-2-Clause",
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/plugin-transform-runtime": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/runtime": "^7.5.5",
+    "@babel/core": "^7.8.4",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/preset-env": "^7.8.4",
+    "@babel/runtime": "^7.8.4",
+    "@osjs/eslint-config": "^1.0.2",
+    "@osjs/stylelint-config": "^1.0.1",
     "babel-loader": "^8.0.6",
-    "copy-webpack-plugin": "^5.0.4",
-    "css-loader": "^2.1.1",
-    "eslint": "^5.16.0",
-    "file-loader": "^3.0.1",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^3.4.2",
+    "eslint": "^6.8.0",
+    "file-loader": "^5.0.2",
     "html-webpack-plugin": "^3.2.0",
-    "hyperapp": "^1.2.10",
-    "mini-css-extract-plugin": "^0.7.0",
-    "node-sass": "^4.12.0",
+    "mini-css-extract-plugin": "^0.9.0",
+    "node-sass": "^4.13.1",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^8.0.2",
     "source-map-loader": "^0.2.4",
-    "stylelint": "^10.1.0",
-    "stylelint-config-standard": "^18.3.0",
-    "stylelint-scss": "^3.9.3",
-    "typeface-roboto": "0.0.54",
-    "webpack": "^4.39.1",
-    "webpack-cli": "^3.3.6"
+    "stylelint": "^13.2.0",
+    "webpack": "^4.41.6",
+    "webpack-cli": "^3.3.11"
   },
   "dependencies": {
     "@osjs/calculator-application": "^1.2.4",
-    "@osjs/cli": "^3.0.18",
-    "@osjs/client": "^3.1.5",
-    "@osjs/dialogs": "^3.0.14",
+    "@osjs/cli": "^3.0.21",
+    "@osjs/client": "^3.1.20",
+    "@osjs/dialogs": "^3.0.17",
     "@osjs/draw-application": "^1.2.3",
-    "@osjs/filemanager-application": "^1.3.4",
+    "@osjs/filemanager-application": "^1.3.9",
     "@osjs/freedesktop-sounds": "^3.0.1",
     "@osjs/gnome-icons": "^4.0.2",
-    "@osjs/gui": "^4.0.25",
+    "@osjs/gui": "^4.0.29",
     "@osjs/htmlviewer-application": "^1.2.3",
     "@osjs/musicplayer-application": "^1.2.5",
-    "@osjs/panels": "^3.0.18",
+    "@osjs/panels": "^3.0.22",
     "@osjs/preview-application": "^1.1.13",
-    "@osjs/server": "^3.0.44",
-    "@osjs/settings-application": "^1.1.1",
-    "@osjs/standard-theme": "^4.0.6",
-    "@osjs/textpad-application": "^1.2.4"
+    "@osjs/server": "^3.0.54",
+    "@osjs/settings-application": "^1.1.2",
+    "@osjs/standard-theme": "^4.0.7",
+    "@osjs/textpad-application": "^1.2.4",
+    "hyperapp": "^1.2.10",
+    "typeface-roboto": "0.0.75"
   },
   "repository": {
     "type": "git",
@@ -78,5 +78,30 @@
   "homepage": "https://www.os-js.org",
   "engines": {
     "node": ">=8.0.0"
+  },
+  "eslintConfig": {
+    "env": {
+      "browser": true,
+      "node": true
+    },
+    "parserOptions": {
+      "sourceType": "module",
+      "ecmaVersion": 8
+    },
+    "extends": "@osjs/eslint-config"
+  },
+  "stylelint": {
+    "extends": "@osjs/stylelint-config"
+  },
+  "babel": {
+    "presets": [
+      [
+        "@babel/preset-env",
+        {}
+      ]
+    ],
+    "plugins": [
+      "@babel/plugin-transform-runtime"
+    ]
   }
 }
