@@ -1,7 +1,7 @@
-/*!
+/*
  * OS.js - JavaScript Cloud/Web Desktop Platform
  *
- * Copyright (c) 2011-2019, Anders Evenrud <andersevenrud@gmail.com>
+ * Copyright (c) 2011-2020, Anders Evenrud <andersevenrud@gmail.com>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -56,14 +56,16 @@ osjs.register(VFSServiceProvider);
 osjs.register(AuthServiceProvider);
 osjs.register(SettingsServiceProvider);
 
-process.on('SIGTERM', () => osjs.destroy());
-process.on('SIGINT', () => osjs.destroy());
-process.on('exit', () => osjs.destroy());
-process.on('uncaughtException', e => console.error(e));
-process.on('unhandledRejection', e => console.error(e));
+const shutdown = signal => (error) => {
+  if (error instanceof Error) {
+    console.error(error);
+  }
 
-osjs.boot()
-  .catch(err => {
-    console.error(err);
-    process.exit(1);
-  });
+  osjs.destroy(() => process.exit(signal));
+};
+
+process.on('SIGTERM', shutdown(0));
+process.on('SIGINT', shutdown(0));
+process.on('exit', shutdown(0));
+
+osjs.boot().catch(shutdown(1));
